requirements.txt
opencv-python
pandas
ultralytics
cvzone
Pillow
numpy
streamlit

#for installation of local packages
-e .
